+++ src/main/java/com/sparta/goods/domain/product/dto/ProductResponse.java
@@ -15,13 +15,13 @@ public class ProductResponse {
 
     private Long productId;
     private String productName;
-    private Double price;
+    private int price;
     private Integer quantity;
     private String introduction;
     private GoodsCategory category;
 
     @Builder
-    public ProductResponse(Long productId, String productName, Double price, Integer quantity,
+    public ProductResponse(Long productId, String productName, int price, Integer quantity,
                            String introduction, GoodsCategory category) {
 
         this.productId = productId;
